+++ client/menu/setup/setup.gd
@@ -37,7 +37,7 @@ func _ready() -> void:
 		if not Network.Client.game_options.empty():
 			$GameOptions.set_game_options(Network.Client.game_options)
 		if not Network.Client.player_list.empty():
-			populate_players(Network.Client.player_list)
+			populate_players({}, Network.Client.player_list)
 			var player_id = multiplayer.get_network_unique_id()
 			if Network.Client.player_list[player_id].spectate == true:
 				$SpectateButton.set_pressed_no_signal(true)
@@ -73,7 +73,7 @@ func set_enable_host_options(is_host: bool) -> void:
 	$IpFinder.visible = ((not Network.Client.is_singleplayer) and is_host)
 
 
-func populate_players(new_player_list: Dictionary) -> void:
+func populate_players(_old_player_list: Dictionary, new_player_list: Dictionary) -> void:
 	Logger.print(self, "Got player list: %s" % [new_player_list])
 	player_list.clear_players()
 	$GameOptions.set_max_bots(Network.MAX_PLAYERS - multiplayer.get_network_connected_peers().size())